coprocessor/local_metrics: use LocalCountVec
