+++ src/coprocessor/dag/dag.rs
@@ -21,12 +21,11 @@ use protobuf::{Message as PbMsg, RepeatedField};
 use coprocessor::codec::mysql;
 use coprocessor::codec::datum::{Datum, DatumEncoder};
 use coprocessor::dag::expr::EvalContext;
-use coprocessor::local_metrics::*;
 use coprocessor::{Error, Result};
 use coprocessor::endpoint::{get_pk, to_pb_error, ReqContext};
 use storage::{Snapshot, SnapshotStore};
 
-use super::executor::{build_exec, Executor, Row};
+use super::executor::{build_exec, Executor, ExecutorMetrics, Row};
 
 pub struct DAGContext {
     columns: Arc<Vec<ColumnInfo>>,
